Added Sessions & Comments to code
I added sessions to my web app, so only when a user is logged in can
they access the web app. Also when they logout the will not be able to
access the web app by URL. I added comments to my python code in the
startPage.py file.
I done some minor changes to the html files, to adjust to the sessions.