Added SQLite3 to project to allow user to add information about Drinks.
I took away some of the html pages and made the web app a single web
page application. I added SQLite to allow the user to input data about
different drinks but SQLite3 is not working properly at the moment and I
can only get the user to input one piece of Data.